Refactor,Add user authentication

+++ MangaLibrary/MangaLibrary.DataAccess/CQRS/Commands/Genre/DeleteGenreCommand.cs
@@ -9,15 +9,16 @@ using System.Threading.Tasks;
 
 namespace MangaLibrary.DataAccess.CQRS.Commands.Genre
 {
-    public class DeleteGenreCommand : CommandBase<Guid, Unit>
+    public class DeleteGenreCommand : CommandBase<Guid, Result<Unit>>
     {
-
-        public override async Task<Unit> Execute(MangaLibraryDbContext context)
+        public override async Task<Result<Unit>> Execute(MangaLibraryDbContext context)
         {
             var item=await context.Genres.FirstOrDefaultAsync(n => n.Id == this.Parameter);
+            if (item == null)
+                return Result<Unit>.Fail($"Genre with id: {this.Parameter} doesn't exist");
             context.Genres.Remove(item);
             await context.SaveChangesAsync();
-            return Unit.Value;
+            return Result<Unit>.Success();  
         }
     }
 }
